+++ src/main/java/net/tlabs/tablesaw/parquet/TablesawRecordConverter.java
@@ -1,4 +1,4 @@
-package tech.tablesaw.io.parquet;
+package net.tlabs.tablesaw.parquet;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -16,6 +16,7 @@ import java.time.temporal.ChronoUnit;
 import java.time.temporal.JulianFields;
 import java.util.List;
 import java.util.Optional;
+import net.tlabs.tablesaw.parquet.TablesawParquetReadOptions.UnnanotatedBinaryAs;
 import org.apache.parquet.Preconditions;
 import org.apache.parquet.io.api.Binary;
 import org.apache.parquet.io.api.Converter;
@@ -50,7 +51,6 @@ import tech.tablesaw.api.Table;
 import tech.tablesaw.api.TextColumn;
 import tech.tablesaw.api.TimeColumn;
 import tech.tablesaw.columns.Column;
-import tech.tablesaw.io.parquet.TablesawParquetReadOptions.UnnanotatedBinaryAs;
 
 public class TablesawRecordConverter extends GroupConverter {
