+++ src/test/java/net/tlabs/tablesaw/parquet/TestMinimizedReadPyArrow.java
@@ -1,32 +1,32 @@
-package tech.tablesaw.io.parquet;
+package net.tlabs.tablesaw.parquet;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
-import java.time.Instant;
+import java.time.LocalDateTime;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.provider.Arguments;
 import tech.tablesaw.api.ColumnType;
 
-class TestDefaultReadFastparquet extends AbstractTableParameterizedTest {
+class TestMinimizedReadPyArrow extends AbstractTableParameterizedTest {
 
   @SuppressWarnings("unused")
   private static Stream<Arguments> columnTypeParameters() {
     return Stream.of(
         Arguments.of(0, ColumnType.BOOLEAN, "boolean"),
-        Arguments.of(1, ColumnType.INTEGER, "uint8"),
-        Arguments.of(2, ColumnType.INTEGER, "int8"),
-        Arguments.of(3, ColumnType.INTEGER, "uint16"),
-        Arguments.of(4, ColumnType.INTEGER, "int16"),
+        Arguments.of(1, ColumnType.SHORT, "uint8"),
+        Arguments.of(2, ColumnType.SHORT, "int8"),
+        Arguments.of(3, ColumnType.SHORT, "uint16"),
+        Arguments.of(4, ColumnType.SHORT, "int16"),
         Arguments.of(5, ColumnType.INTEGER, "uint32"),
         Arguments.of(6, ColumnType.INTEGER, "int32"),
         Arguments.of(7, ColumnType.LONG, "uint64"),
         Arguments.of(8, ColumnType.LONG, "int64"),
-        Arguments.of(9, ColumnType.DOUBLE, "float32"),
+        Arguments.of(9, ColumnType.FLOAT, "float32"),
         Arguments.of(10, ColumnType.DOUBLE, "float64"),
-        Arguments.of(11, ColumnType.INSTANT, "TIMESTAMP_MICROS"),
+        Arguments.of(11, ColumnType.LOCAL_DATE_TIME, "TIMESTAMP_MILLIS"),
         Arguments.of(12, ColumnType.STRING, "string"));
   }
 
@@ -35,14 +35,14 @@ class TestDefaultReadFastparquet extends AbstractTableParameterizedTest {
     return Stream.of(
         Arguments.of(0, 0, true, "boolean"),
         Arguments.of(0, 1, false, "boolean"),
-        Arguments.of(1, 0, 0, "integer"),
-        Arguments.of(1, 1, 127, "integer"),
-        Arguments.of(2, 0, -127, "integer"),
-        Arguments.of(2, 1, 1, "integer"),
-        Arguments.of(3, 0, 0, "integer"),
-        Arguments.of(3, 1, 32767, "integer"),
-        Arguments.of(4, 0, 0, "integer"),
-        Arguments.of(4, 1, -32767, "integer"),
+        Arguments.of(1, 0, (short) 0, "short"),
+        Arguments.of(1, 1, (short) 127, "short"),
+        Arguments.of(2, 0, (short) -127, "short"),
+        Arguments.of(2, 1, (short) 1, "short"),
+        Arguments.of(3, 0, (short) 0, "short"),
+        Arguments.of(3, 1, (short) 32767, "short"),
+        Arguments.of(4, 0, (short) 0, "short"),
+        Arguments.of(4, 1, (short) -32767, "short"),
         Arguments.of(5, 0, 0, "integer"),
         Arguments.of(5, 1, 65000, "integer"),
         Arguments.of(6, 0, 0, "integer"),
@@ -51,23 +51,26 @@ class TestDefaultReadFastparquet extends AbstractTableParameterizedTest {
         Arguments.of(7, 1, 1_000_000_000L, "long"),
         Arguments.of(8, 0, 0L, "long"),
         Arguments.of(8, 1, -1_000_000_000L, "long"),
-        Arguments.of(9, 0, null, "double"),
-        Arguments.of(9, 1, 1.0d, "double"),
+        Arguments.of(9, 0, null, "float"),
+        Arguments.of(9, 1, 1.0f, "float"),
         Arguments.of(10, 0, 0.0d, "double"),
         Arguments.of(10, 1, null, "double"),
-        Arguments.of(11, 0, Instant.parse("2021-04-23T00:00:00Z"), "Instant"),
-        Arguments.of(11, 1, Instant.parse("2021-04-23T00:00:01Z"), "Instant"),
+        Arguments.of(11, 0, LocalDateTime.of(2021, 4, 23, 0, 0), "LocalDateTime"),
+        Arguments.of(11, 1, LocalDateTime.of(2021, 4, 23, 0, 0, 1), "LocalDateTime"),
         Arguments.of(12, 0, "string1", "String"),
         Arguments.of(12, 1, "string2", "String"));
   }
 
-  private static final String PANDAS_FASTPARQUET = "target/test-classes/pandas_fastparquet.parquet";
+  private static final String PANDAS_PYARROW = "target/test-classes/pandas_pyarrow.parquet";
 
   @BeforeAll
   static void beforeAll() throws IOException {
     table =
         new TablesawParquetReader()
-            .read(TablesawParquetReadOptions.builder(new File(PANDAS_FASTPARQUET)).build());
-    assertEquals("pandas_fastparquet.parquet", table.name(), "Wrong table name");
+            .read(
+                TablesawParquetReadOptions.builder(new File(PANDAS_PYARROW))
+                    .minimizeColumnSizes()
+                    .build());
+    assertEquals("pandas_pyarrow.parquet", table.name(), "Wrong table name");
   }
 }
